requirements.txt
pandas
numpy 
pymongo
PyYAML
python-dotenv
dill
dnspython
fastapi
httptools
imblearn
mypy-boto3-s3
pip-chill
python-dotenv
types-s3transfer
uvicorn
watchfiles
websockets
wincertstore
xgboost
-e .

